# swagger-2.0.yaml
swagger: '2.0'
info:
  title: Authentication API
  description: |
    Swagger 2.0 specification for the authentication-api service

    ## API Response Codes

    ### Success Codes
    | Code  | Message |
    |-------|---------|
    | SUC000 | Success |
    | SUC001 | Created successfully |
    | SUC002 | Updated successfully |
    | SUC003 | Deleted successfully |
    | SUC004 | Fetched successfully |

    ### Error Codes
    | Code  | Message |
    |-------|---------|
    | ERR000 | Invalid request |
    | ERR001 | Internal server error |
    | ERR002 | User not found |
    | ERR003 | User already exists |
    | ERR004 | Invalid credentials |
    | ERR005 | Invalid token |
    | ERR006 | Expired token |
    | ERR007 | Unauthorized |
  version: '1.0.0'
host: localhost:3000
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json

x-api-response-codes:
  # Success Codes
  SUC000: Success
  SUC001: Created successfully
  SUC002: Updated successfully
  SUC003: Deleted successfully
  SUC004: Fetched successfully
  # Error Codes
  ERR000: Invalid request
  ERR001: Internal server error
  ERR002: User not found
  ERR003: User already exists
  ERR004: Invalid credentials
  ERR005: Invalid token
  ERR006: Expired token
  ERR007: Unauthorized

paths:
  /:
    get:
      tags:
        - Health
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: Successful response with greeting text.
          schema:
            type: string
          examples:
            application/json: Hello World!

  /auth/signin:
    post:
      tags:
        - Auth
      summary: Sign in user
      operationId: signIn
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SignInRequest'
      responses:
        '200':
          description: Successful sign-in response.
          schema:
            type: object

  /auth/signup:
    post:
      tags:
        - Auth
      summary: Sign up new user
      operationId: signUp
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SignUpRequest'
      responses:
        '201':
          description: User created successfully.
          schema:
            type: object

  /auth/signout:
    post:
      tags:
        - Auth
      summary: Sign out user
      operationId: signOut
      responses:
        '200':
          description: Successful sign-out response.
          schema:
            type: object

  /auth/refresh-token:
    post:
      tags:
        - Auth
      summary: Refresh access token
      operationId: refreshToken
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/TokenRequest'
      responses:
        '200':
          description: Token refreshed successfully.
          schema:
            type: object

  /auth/send-recovery-code:
    post:
      tags:
        - Auth
      summary: Send password recovery code to email
      operationId: sendRecoveryCode
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/EmailRequest'
      responses:
        '200':
          description: Recovery code sent successfully.
          schema:
            type: object

  /auth/verify-recovery-code:
    post:
      tags:
        - Auth
      summary: Verify password recovery code
      operationId: verifyRecoveryCode
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/VerifyRecoveryCodeRequest'
      responses:
        '200':
          description: Recovery code verified successfully.
          schema:
            type: object

  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset user password
      operationId: resetPassword
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully.
          schema:
            type: object

  /auth/google:
    post:
      tags:
        - Auth
      summary: Sign in with Google
      operationId: googleSignIn
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/TokenRequest'
      responses:
        '200':
          description: Successful Google sign-in response.
          schema:
            type: object

  /users:
    get:
      tags:
        - Users
      summary: Get all users
      operationId: findAllUsers
      responses:
        '200':
          description: List of users.
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createUser
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateUserRequest'
      responses:
        '201':
          description: User created successfully.
          schema:
            $ref: '#/definitions/User'

  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      operationId: findOneUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        '200':
          description: User found.
          schema:
            $ref: '#/definitions/User'
    put:
      tags:
        - Users
      summary: Update user by ID
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateUserRequest'
      responses:
        '200':
          description: User updated successfully.
          schema:
            $ref: '#/definitions/User'
    patch:
      tags:
        - Users
      summary: Partially update user by ID
      operationId: patchUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateUserRequest'
      responses:
        '200':
          description: User patched successfully.
          schema:
            $ref: '#/definitions/User'
    delete:
      tags:
        - Users
      summary: Delete user by ID
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        '200':
          description: User deleted successfully.
          schema:
            $ref: '#/definitions/DeleteResponse'